Added traits check (profile photo and API) when deleting a user to prevent tests to fail

+++ stubs/app/Actions/Jetstream/DeleteUser.php
@@ -3,6 +3,8 @@
 namespace App\Actions\Jetstream;
 
 use Laravel\Jetstream\Contracts\DeletesUsers;
+use Laravel\Jetstream\HasProfilePhoto;
+use Laravel\Sanctum\Contracts\HasApiTokens;
 
 class DeleteUser implements DeletesUsers
 {
@@ -14,8 +16,17 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
-        $user->deleteProfilePhoto();
-        $user->tokens->each->delete();
+    
+        $userModelTraits = class_uses($user);
+    
+        if (in_array(HasProfilePhoto::class, $userModelTraits)) {
+            $user->deleteProfilePhoto();
+        }
+    
+        if (in_array(HasApiTokens::class, $userModelTraits)) {
+            $user->tokens->each->delete();
+        }
+    
         $user->delete();
     }
 }
